int changed with long

+++ APBDFinalProject/APBDFinalProject/Models/RequestModels/IndividualCustomerRequest.cs
@@ -2,7 +2,7 @@ namespace APBDFinalProject.RequestModels;
 
 public class IndividualCustomerRequest
 {
-    public int PESEL { get; set; }
+    public long PESEL { get; set; }
     public string FirstName { get; set; }
     public string LastName { get; set; }
     public string Address { get; set; }
